move ERC20 Address and Hash to root types dir
