+++ Jumpscale/clients/blockchain/tfchain/types/transactions/ERC20.py
@@ -4,56 +4,10 @@ from .Base import TransactionBaseClass, TransactionVersion, InputSignatureHashFa
 
 from ..FulfillmentTypes import ED25519Signature, SignatureCallbackBase, SignatureRequest
 from ..ConditionTypes import UnlockHash
-from ..PrimitiveTypes import BinaryData, Hash, Currency
+from ..PrimitiveTypes import BinaryData, Currency
+from ..ERC20 import ERC20Address, ERC20Hash
 from ..IO import CoinInput, CoinOutput
 from ..CryptoTypes import PublicKey
-
-from abc import abstractmethod
-
-class ERC20Address(BinaryData):
-    SIZE = 20
-
-    """
-    ERC20 Contract/Wallet Address, used in TFChain-ERC20 code.
-    """
-    def __init__(self, value=None):
-        super().__init__(value, fixed_size=ERC20Address.SIZE, strencoding='hexprefix')
-
-    @classmethod
-    def from_unlockhash(cls, unlockhash):
-        """
-        Create an ERC20 Address from a TFT Address (type: UnlockHash).
-        """
-        e = j.data.rivine.encoder_sia_get()
-        unlockhash.sia_binary_encode(e)
-        hash = bytes.fromhex(j.data.hash.blake2_string(e.data))
-        return cls(value=hash[Hash.SIZE-ERC20Address.SIZE:])
-
-    @classmethod
-    def from_json(cls, obj):
-        if obj is not None and not isinstance(obj, str):
-            raise TypeError("ERC20 address is expected to be an encoded string when part of a JSON object, not {}".format(type(obj)))
-        if obj == '':
-            obj = None
-        return cls(value=obj)
-
-class ERC20Hash(BinaryData):
-    SIZE = 32
-
-    """
-    ERC20 Hash, used in TFChain-ERC20 code.
-    """
-    def __init__(self, value=None):
-        super().__init__(value, fixed_size=ERC20Hash.SIZE, strencoding='hexprefix')
-
-    @classmethod
-    def from_json(cls, obj):
-        if obj is not None and not isinstance(obj, str):
-            raise TypeError("ERC20 hash is expected to be an encoded string when part of a JSON object, not {}".format(type(obj)))
-        if obj == '':
-            obj = None
-        return cls(value=obj)
-
 
 class TransactionV208(TransactionBaseClass):
     _SPECIFIER = b'erc20 convert tx'
